fix excel

+++ server/database/db/leave_request/20190925/alter_table.sql
@@ -1,3 +1,3 @@
 ALTER TABLE leave_request 
-ADD COLUMN before_leave_balance int,
-ADD COLUMN after_leave_balance int;
+ADD COLUMN before_leave_balance float,
+ADD COLUMN after_leave_balance float;
